--- include/exception/ArgumentParseException.hpp
@@ -1,5 +1,0 @@
-#ifndef FT_IRC_ARGUMENT_PARSE_EXCEPTION
-#define FT_IRC_ARGUMENT_PARSE_EXCEPTION
-
-
-#endif /* FT_IRC_ARGUMENT_PARSE_EXCEPTION */
