+++ modules/core/includes/exception/ArgumentParseException.hpp
@@ -1,0 +1,5 @@
+#ifndef FT_IRC_CORE_ARGUMENT_PARSE_EXCEPTION
+#define FT_IRC_CORE_ARGUMENT_PARSE_EXCEPTION
+
+
+#endif /* FT_IRC_ARGUMENT_PARSE_EXCEPTION */
