+++ modules/core/includes/exception/CommandNotFoundException.hpp
