+++ modules/parsing/includes/CommandExecutor.hpp
@@ -1,7 +1,7 @@
 #ifndef FT_IRC_COMMAND_EXECUTOR
 #define FT_IRC_COMMAND_EXECUTOR
 
-#include "core.hpp"
+#include "ft_irc.hpp"
 #include "Response.hpp"
 #include "command/Command.hpp"
 #include "command/CommandSender.hpp"
